Add permission statement
The LICENSE file contains the text of the GPLv2 and also instructions
on how to release a program under "GPLv2 or (at your option) any later
version".  However even though that file contains instructions on how
to move from "GPLv2 and only v2" to "GPLv2 or later" one has to
actually follow those instructions.

In issue #7 Javier Olaechea asked for help adding the "... or later".
This is what I am doing with this commit.